Code style improvement; self-hosted runner

+++ tests/test_notion.py
@@ -4,20 +4,15 @@ Test all interactions with notion from each parser events.
 import os
 from datetime import datetime
 
+import psycopg2
 import pytest
 
-from davai_s_nami_bot import database, events, notion_api, telegram, vk
+from davai_s_nami_bot import database, events, notion_api, telegram
 from davai_s_nami_bot.notion_api import notion_client
 
-test_table1 = notion_client.get_collection_view(
-    os.environ.get("NOTION_TEST_TABLE1_URL")
-)
-test_table2 = notion_client.get_collection_view(
-    os.environ.get("NOTION_TEST_TABLE2_URL")
-)
-test_table3 = notion_client.get_collection_view(
-    os.environ.get("NOTION_TEST_TABLE3_URL")
-)
+test_table1 = notion_client.get_collection_view(os.environ.get("NOTION_TEST_TABLE1_URL"))
+test_table2 = notion_client.get_collection_view(os.environ.get("NOTION_TEST_TABLE2_URL"))
+test_table3 = notion_client.get_collection_view(os.environ.get("NOTION_TEST_TABLE3_URL"))
 
 
 def move_rows(from_table, to_table):
@@ -69,20 +64,26 @@ def test_move_from_dev_table1_to_dev_table3():
 
 def test_post_event_from_dev_table3():
     """
-    Required events in dev table 3 (from test 'test_move_from_dev_table_2_to_dev_table_3'
-    or 'test_move_from_dev_table_1_to_dev_table_3').
+    Required events in dev table 3 (from test
+    'test_move_from_dev_table_2_to_dev_table_3' or
+    'test_move_from_dev_table_1_to_dev_table_3').
     """
     for row in test_table3.collection.get_rows():
         event = notion_api.notion_row_to_event(row)
 
-        telegram.send_post(event)
+        if event.Event_id in database.get_all()["id"]:
+            with pytest.raises(psycopg2.errors.UniqueViolation):
+                telegram.send_post(event)
 
-        database.remove_by_event_id(event.Event_id)
+        else:
+            telegram.send_post(event)
+            database.remove_by_event_id(event.Event_id)
 
 
 def test_move_from_dev_table2_to_dev_table3():
     """
-    Required events in dev table 2 (from test 'test_move_from_dev_table_1_to_dev_table2').
+    Required events in dev table 2 (from test
+    'test_move_from_dev_table_1_to_dev_table2').
     """
     move_rows(test_table2, test_table3)
